+++ tests/test_mosaic.py
@@ -748,7 +748,7 @@ def test_mosaic_list(app):
     resp = response.json()
     assert resp["context"]
     dates = [
-        datetime.strptime(s["search"]["lastused"][0:-6], "%Y-%m-%dT%H:%M:%S.%f")
+        datetime.strptime(s["search"]["lastused"], "%Y-%m-%dT%H:%M:%S.%fZ")
         for s in resp["searches"]
     ]
     assert dates[0] < dates[-1]
@@ -758,7 +758,7 @@ def test_mosaic_list(app):
     resp = response.json()
     assert resp["context"]
     dates = [
-        datetime.strptime(s["search"]["lastused"][0:-6], "%Y-%m-%dT%H:%M:%S.%f")
+        datetime.strptime(s["search"]["lastused"], "%Y-%m-%dT%H:%M:%S.%fZ")
         for s in resp["searches"]
     ]
     assert dates[0] > dates[-1]