+++ titiler/pgstac/dependencies.py
@@ -1,6 +1,5 @@
 """titiler-pgstac dependencies."""
 
-import sys
 from dataclasses import dataclass, field
 from typing import Optional, Tuple
 
@@ -13,17 +12,12 @@ from psycopg import errors as pgErrors
 from psycopg.rows import dict_row
 from psycopg_pool import ConnectionPool
 from starlette.requests import Request
+from typing_extensions import Annotated
 
 from titiler.core.dependencies import DefaultDependency
 from titiler.pgstac import model
 from titiler.pgstac.settings import CacheSettings, RetrySettings
 from titiler.pgstac.utils import retry
-
-if sys.version_info >= (3, 9):
-    from typing import Annotated  # pylint: disable=no-name-in-module
-else:
-    from typing_extensions import Annotated
-
 
 cache_config = CacheSettings()
 retry_config = RetrySettings()
@@ -43,7 +37,7 @@ def SearchParams(
     body: model.RegisterMosaic,
 ) -> Tuple[model.PgSTACSearch, model.Metadata]:
     """Search parameters."""
-    search = body.dict(
+    search = body.model_dump(
         exclude_none=True,
         exclude={"metadata"},
         by_alias=True,
@@ -121,7 +115,7 @@ def get_stac_item(pool: ConnectionPool, collection: str, item: str) -> pystac.It
         with conn.cursor(row_factory=dict_row) as cursor:
             cursor.execute(
                 ("SELECT * FROM pgstac.search(%s) LIMIT 1;"),
-                (search.json(by_alias=True, exclude_none=True),),
+                (search.model_dump_json(by_alias=True, exclude_none=True),),
             )
 
             resp = cursor.fetchone()["search"]